+++ Level4-Module2/src/_03_generics_store/_02_food/Cereal.java
@@ -1,16 +1,16 @@
-package _03_generics_store;
+package _03_generics_store._02_food;
 
 import java.util.Random;
 
 import javax.swing.JLabel;
 
-public class Candy extends Food{
-	public Candy() {
+public class Cereal extends Food{
+	public Cereal() {
 		if(new Random().nextBoolean()) {
-			this.item = "images/candy1.jpeg";
+			this.item = "images/cereal1.jpeg";
 		}
 		else {
-			this.item = "images/candy2.jpeg";
+			this.item = "images/cereal2.jpeg";
 		}
 	}
